Upgraded OL3 version to 3.11.2
Upgraded the version of ol3 the application uses to the latest version.
This version has added client side raster re-projection.  this can be
used  in the application by setting a different coord system in the "Map
Options", ol3 will then do the rest.  results are variable and have high
client side processing requirements outside of coord bounds....... use
with caution.